+++ src/Resources/contao/config/config.php
@@ -11,6 +11,11 @@
  * @copyright Frank Hoppe 2014
  */
 
+if(TL_MODE == 'FE' && isset($GLOBALS['TL_CONFIG']['teamtournament_css']))
+{
+	$GLOBALS['TL_CSS'][] = 'bundles/contaoteamtournament/default.css';
+}
+
 $GLOBALS['BE_MOD']['content']['teamtournament'] = array
 (
 	'tables'         => array('tl_teamtournament', 'tl_teamtournament_teams', 'tl_teamtournament_players', 'tl_teamtournament_matches', 'tl_teamtournament_games'),