+++ src/Resources/contao/dca/tl_content.php
@@ -51,7 +51,7 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['teamtournament_runde'] = array
 (
 	'label'                  => &$GLOBALS['TL_LANG']['tl_content']['teamtournament_runde'],
 	'exclude'                => true,
-	'options'                => array(1, 2, 3, 4, 5, 6, 7, 8, 9),
+	'options'                => array(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19),
 	'inputType'              => 'select',
 	'eval'                   => array
 	(