+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,5 +1,8 @@
 <?php 
 
 $GLOBALS['TL_LANG']['tl_content']['tt-lineup_legend'] = 'Einstellungen';
-
 $GLOBALS['TL_LANG']['tl_content']['teamtournament_lineup'] = array('Mannschaft', 'Mannschaft auswählen');
+
+$GLOBALS['TL_LANG']['tl_content']['tt-pairings_legend'] = 'Turnier und Runde';
+$GLOBALS['TL_LANG']['tl_content']['teamtournament_turnier'] = array('Turnier', 'Turnier auswählen');
+$GLOBALS['TL_LANG']['tl_content']['teamtournament_runde'] = array('Runde', 'Runde auswählen');